Use wiki as a database to persistent build id between runs
Artifacts are only persistent in the action during a workflow run.
Also, it’s sensitive to a decaying period where the unused artifact
will be cleaned up.
Consequently, rather than storing it permanently here, store it in the
wiki where we can edit it.
We join all results in a single jobs, once the matrix jobs completes.

Also, don’t fail the matrix when one of them is failing. Keep the other
platforms updated and uploading to the store. We will then only
collected fully completed workflows.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>