Bringing not_null<> to the project.
There are situations when holding a pointer as a member results in
simpler code than holding a reference type.
Yet this causes the loss of the guarantee that references gives us of
never being null.
This template type was copied from Microsoft's implementation of the
Core Guidelines Support Library (MIT License).
not_null<> is a pointer that won't compile if initialized with nullptr.

See https://github.com/microsoft/GSL