The bare bones state machine
This generic state machine component will be the foundation for the
splash controller and other components with a similar role/philosophy.
It makes extensive use of C++ templates with some meta programming to
enable a generic but sanitized usage of std::variant.
Instead of the classic GoF State design pattern with class hierarchies,
heavy use of pointers, virtual methods and double-dispatch, this state
machine builds upon the value semantics of sum types.
std::variant can be seen as a discriminated union, thus it's type safe.
Using some template magic we can build a generic component that models
states and events as variants, handles them exaustively and makes sure
to transport the invalid state transitions appropriately.
There are nicer state machine implementations out there, even with nicer
compile-time features, like refusing to compile an invalid transition.
That would require adding dependencies, what we are attempting to avoid
for this cycle.
Although this state machine may see a bit scary, its just a variant of a
pattern that has been presented in conferences for the past few years,
at least since standardization of std::variant in C++17.

A few examples (this is heavily based on the ideas of the first talk)
- CppCon 2018: Mateusz Pusz “Effective replacement of dynamic polymorphism with std::variant”: https://www.youtube.com/watch?v=gKbORJtnVu8
- Pacific++ 2017: Nick Sarten "Type-safe state machines with C++17 std::variant": https://www.youtube.com/watch?v=4v-pPkVaFA4
- std variant and the power of pattern matching - Nikolai Wuttke - Meeting C++ 2018: https://www.youtube.com/watch?v=CELWr9roNno