Added a Strategy to abstract out OS interaction
I hope the demonstration in SplashControllerTests.cpp convinces about
the indirection added. By abstracting out the interactions with the
operating system, we can reason about the state-event logic more easily
and assert many situations in automated tests that would be hard to do
otherwise. I had to experiment controlling this splash application in
many ways to decide that the apparent complexity added with those state
machines and strategy classes are worth. Specially because we loose some
readability with the addition of the template syntax. The test flexibility
convinced me that it was worth trying out.