Fix tests and remove forcebly close.
As a consequence of last modifications, the preferred way to move the attention from the splash to the installer GUI is by hiding the splash instead of closing it.
TUI is still handled by closing the splash window.
Flutter subproject was already adjusted for that. See PR #8.
The do_forcebly_close function lost significance and is being removed.