PackageDependency VCLibs.140.00.UWPDesktop
According to the following documentation, adding the PackageDependency
entry to the project ensures automatic installation of the VC++ runtime
on hosts where it's missing:

https://docs.microsoft.com/en-us/troubleshoot/developer/visualstudio/cpp/libraries/c-runtime-packages-desktop-bridge

This is helpful for the Flutter splash screen, but might be for the launcher itself as well.

More information about this entry:

https://docs.microsoft.com/en-us/uwp/schemas/appxpackage/uapmanifestschema/element-packagedependency

Apparently this is not testable with sideloading, so we can only be sure
after having a appx deployed to the store and tested on a proper target.
It has to be a non-development machine, like a win10-wsl vm we have on azure.