Fix the distro reboot behavior.
Previously we had one extra run with root.
Now we configure the default user first, then log with it if we get to the shell.
There is no proper shutdown for this case, because if it was not install only,
then the upstream code will run the shell with the default logged user.