Use specific template files for project files
In order to make the project always buildable, from Visual Studio for
example, project files must not be templated. Instead of that we create
a separate file with a .tmpl extension which is the actual template.
This file will be parsed and copied to the destination without the .tmpl
extension. The template tags are removed from the original project
files. Be careful though because the original project files are still
parsed and if they contain a template tag, they'll overwrite the real
template.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>