addEvent easier to compile with r-value ref.
This small change should reduce the effort the compiler does to
differentiate the preferred case of addEvent overloaded selection
against the general case (auto&&).
Compiler never complained about this, the linter (clang-tidy) was
crashing badly during instantiation of this template, which leads to an
educated guess that internally it was having trouble to differentiate
between the cases to select which overload would be preferred.
Removing one '&' turns more obvious the preferrence, since the visit
function will cause the variant extraction to generate an R-value
reference, which in turn binds more obviously to '&' than '&&'.
Educated guess. I don't see a practical way to measure the complexity
of the compilation from the point of view of the linter.
Fact is that removing one '&' from the preferred cases stopped crash.