Brute force run clang-tidy.
Needs to investigate and bug report.
Same crash is happening without clear indication of cause.
On the same template instantiation. Could be a bug in the tool.
Either way, when it happens a huge exit code number is generated.
Local tests revealed that try again could pass thru the error.
Will see in the CI.