Try fixing dailies with no development releases
As we are in a particular state with no release to build at all, then
the slice of matrixEleme is deserialized to null, which is json
compatible.
It seems that GitHub Actions don’t like that and still expect an
iterable.

Consequently, this is a tentative fix to set it to an empty slice
instead of nil, so that GitHub Action could try iterating over it and do
nothing.