Common bits of the App strategies
Simpler strategies that lack the interlocking between the installer and the slide show benefit from those implementations.
When migrating the OOBE to Windows a new strategy very close to the NoSplashStrategy implemented for ARM64, but not exactly equal, would cause a lot of code duplication if those were not extracted.

This refactoring won't change the current behavior.