Allows creating a new console for the child process ...
if appConfig requires so.

Apparently we need that to be able to run the Go application.
I really would like to avoid that.
We are accessing this appConfig() in a very low level.
I hope to soon figure out how to avoid this.
It will be a pleasure to revert this commit.