Initial workflow implementation
This workflow file is capable of running WSL from Store.
That cannot be done from session 0.
An Ubuntu instance is raised just to RDP into the Azure runner.
The Azure runner then uses SysInternals PsExec to execute Powershell
into the RDP Logon session (typically 2).

The Command passed to powershell will then be replaced by the Go Test.
Before that we need to install Go, Visual Studio Build Tools, and
invoke the preparation script to build and install the appx.