Prevents Ubuntu and Ubuntu-22.04 installation problem
The failure was a very strange complain about not being able to merge a resource.pri file
from some inner asset package of the bundle.
Removing the conditional inclusion of the Flutter app prevented the error.
The relationship between the problem and the fix is not apparent.
Also, the Ubuntu-Preview has the same conditional but installs correctly.
Further troubleshooting will be required later.
For now let's get the problem sorted out.