Addresses loss of comm between splash and launcher
The wait introduced in https://github.com/ubuntu/WSL/pull/278 broke comm between splash and launcher
Splash communicates its HWND on creation.
That requires the launcher to have already created the named pipe.
The wait caused the splash window to try to communicate before the named pipe was created.
That wait is only useful if the child is a console app (the case for E2E).
Thus the appConfig check makes sense.