Avoid triggering Refresh releases and assets
Refresh releases and assets is an automated action that looks for the string "UbuntuDev.WslID.Dev". If this string is found, then multiple copies of the file are generated in which this string is replaced with "Ubuntu-18.08", "Ubuntu-20.04", etc. To avoid triggering this, I changed the code so it does not contain string "UbuntuDev.WslID.Dev".