(fix) Cannot read output from previous job steps
Basic GitHub actions mistake.
We are attempting to store the debug symbols of the latest published
apps in the wiki.
For that we need to know whether this build lead to a new submission to
the store.
We were attempting to read the output from a previous job step.
That can only work if the job itself makes the output available.