Machinery to patch config files after registration
Implements a PatchConfig that declares a file and a patching function to act upon it.
And a Patcher to segregate messing with files from the actual patch logic.
A higher level construct will use or create a collection of PatchConfigs and instantiate Patchers to apply them.

The highest level construct of this set of changes it the
`apply(PatchConfig, prefixPath)` function.

It takes care of instantiating a Patcher and executing the callback.
The path prefix is needed by the Patcher to know how to access the Linux files from Windows.

It relies a lot on C++17 filkesystem API.
Success can only be guaranteed if root access is ensured.
Thus the reason for only using it upon distro registration.