Implements the means of disabling units
via systemctl disable.
Handles cases when patching config files is not an option.
For now everything hardcoded.
We only have one unit in one release to be disabled.

If this evolves, something similar to the patch mechanism
should be used to select the units per release and compute the
systemctl command line.