Fix obscure Git error in wsl-checkout
Apparently you cannot fetch the branch you're in. This is not an issue
when this action is called in a PR:
- `git clone ...` clones into main
- `git fetch ...` fetches the PR ref

However, when this action is ran on main (for instance, after merging
the PR), the fetch will try to fetch main, and fail.

See here:
https://github.com/ubuntu/wsl-actions-example/actions/runs/5120947427/jobs/9208564434

See a Stackoverflow thread about this:
https://stackoverflow.com/questions/2236743/git-refusing-to-fetch-into-current-branch