feat: Make the launcher aware of cloud-init
This affects only the registration time.
Subsequent boots don't check for cloud-init in any way.
We check if there is a `cloud-init` command in PATH,
and launch `cloud-init --status` right after registration is complete.
This will wait until cloud-init completes what's doing or report
error/disabled states.

Cloud-init could have created Linux user accounts.
We expect that it sets the desired default one via `/etc/wsl.conf`.
That way the effects would show up even without the launcher.
We parse that file and, if there is the default user entry, then
we set it via WSL API (which is faster than terminanting and restarting
the instace).

If we don't find a default user set, than we fallback to the interative
user creation, i.e. the default upstream prompting.