feat: Make the launcher aware of cloud-init
This affects only the registration time.
Subsequent boots don't check for cloud-init in any way.
We check if there is a `cloud-init` command in PATH,
and launch `cloud-init --status` right after registration is complete.
This will wait until cloud-init completes what's doing or report
error/disabled states.

Cloud-init could have created Linux user accounts.
We expect that it sets the desired default one via `/etc/wsl.conf`.
That way the effects would show up even without the launcher.
We parse that file and, if there is the default user entry, then
we set it via WSL API (which is faster than terminanting and restarting
the instance).

If we don't find a default user set, than we fallback to the interative
user creation, i.e. the default upstream prompting.

I originally intended to put this implementation inside
DistributionInfo.h/cpp to ensure less files being touched (thus less
diff compared to upstream) but that doesn't play nice with CI 'Refresh
releases and assets' workflow.

So I placed the code in its own component, and inside a subdirectory,
so if there comes a time we need to revert that change (it becomes
obsolete or so) it remains easy to drop.
Also, being inside a subdirectory allows me to drop a .clang-format in
there without worrying of reformatting upstream code :)

Lessons learned from the launcher old times:
- We can put cpp files inside a subdirectory and still work with
  precompiled headers in Visual Studio if we either adjust the include
  path for every cpp file or adjust the path to the precompiled header
  for every cpp file (no winner :) )