feat: Make the launcher aware of cloud-init
This affects only the registration time.
Subsequent boots don't check for cloud-init in any way.

We check if there is a `cloud-init` command in PATH,
and launch `cloud-init status --wait` right after registration is complete.
This will wait until cloud-init completes what's doing or report
error/disabled states.

Cloud-init could have created Linux user accounts. Users could have been
created in many other ways, even pre-created during image build.
We expect the desired default user to be set via `/etc/wsl.conf`.
That would work even without the distro launcher, so it's reasonable to
assume the desired default to be present there.

We "parse" that file and, if there is the default user entry, then
we set it via WSL API (which is faster than terminanting and restarting
the instance).
When "parsing" /etc/wsl.conf (which is INI like syntax), we deliberately
stop on the first occurence of the [section].key we're looking for.
That matches the behavior observed in WSL itself.

If we don't find a default user set, we check if the current default
user is still root, in which case we look for any non-system
user in the NSS passwd database (by launching getent),
i.e. any user with UID between 1000 and 65534 (the nobody).
Finally, we fallback to the interative user creation,
i.e. the default upstream prompting.

We do some effort to keep /etc/wsl.conf and the Windows registry (via
WSL API) in sync, in the sense that whatever user we set as default via
one means we also set in the other.

I originally intended to put this implementation inside
DistributionInfo.h/cpp to ensure less files being touched (thus less
diff compared to upstream) but that doesn't play nice with CI 'Refresh
releases and assets' workflow.

So I placed the code in its own component, and inside a subdirectory,
so if there comes a time we need to revert that change (it becomes
obsolete or so) it remains easy to drop.
Also, being inside a subdirectory allows me to drop a .clang-format in
there without worrying of reformatting upstream code :)

Lessons learned from the launcher old times:
- We can put cpp files inside a subdirectory and still work with
  precompiled headers in Visual Studio if we either adjust the include
  path for every cpp file or adjust the path to the precompiled header
  for every cpp file (no winner :) )