Implements tests that rely on cloud-init
Every testcase wipes out the WSL instance
So not a lot can run in a reasonable way in CI
Let's see how long those two will take.

We use `<launcher.exe> install` to avoid the interactive shell in the end.

If we append `--root` then the launcher should still wait for cloud-init but skip checking the creation of the default user.