Fix testCorrectUpgradePolicy
It assumed distroName to match the AppID
Which is not the case.
Ubuntu24.04LTS is the AppID for the distroName Ubuntu-24.04.
Thus we need to check for Ubuntu-WX.YZ :)

This is yet a bit tricky because nowadays we set the policy during image build,
So a rootfs meant to build an LTS app (thus Prompt=never)
could be recycled to build Ubuntu-Preview during tests and that would break the assertion.

TODO: Verify if CI is happy with just this change.