Replace custom ini "parser" by GetPrivateProfileString
With a caveat: that function can only read local files, i.e.
in the Windows OS filesystem.
No shared mounts.
Thus we make an effort to copy the instance's /etc/wsl.conf
to a private (best effort) temporary path, using OS primitives to guaratee automatic deletion.
From that place, it's just a function call to read the key we are interested in.
Let's trust the OS can deliver us a suitable temporary file under %TEMP%.
In production that's more likely, as the %TEMP% dir is private to the app when FS virtualization is ON.