The most beautiful C++ passwd parser ever written ;)
Relying only on the UID may be misleading
So we wanted to check the login shell (the last field of the line) to determine
whether an account is a system or real user.

That required parsing the passwd more carefully
Using istringstreams as before would incur in potentially lots of small memory allocations.
Writing an iteration scheme by hand with string_views prevents most of the unnecessary allocations
while offering an elegant, idiomatic iteration and parsing on-the-flight.

Most of that code wouldn't be needed in C++20, std::ranges::split would do most of this job,
but with C++17 we need to craft the iteration by hand.