[NEW!!!] Adds test cases for higher UIDs and broken passwd
We could break passwd in one way at a time,
but I'm affraid of having tons of test cases making CI run for too long.
So, as the system happily ignores broken lines, I'm all for breaking them in all the ways we want in the same tests
and have few tests running with broken passwd.