Update cloud-init tutorial for 24.04.1
With that release comes changes to the distro launcher making no longer
necessary for users to run cloud-init commands. The distro launcher does
it, so the tutorial can be made shorter.

Also there was a caveat about `wsl --install` that's no longer
applicable, as that installation command also goes through the launcher.