Turn off the WSL1 test case for now.
The VM was crashing. There are quite a few events like the one below that happened yesterday and today.

![image](https://github.com/user-attachments/assets/f6160c34-08e1-4fdd-9b3d-9f79a76001e3)

I resized it to the [second VM size compatible with GitHub large Windows runners](https://docs.github.com/en/actions/using-github-hosted-runners/using-larger-runners/about-larger-runners#specifications-for-general-larger-runners) (which are available by default now for open source projects) so that, if we were to migrate this out of our self-hosted runners we wouldn't be surprised with a downgrade. I then resized to the first size, as the change didn't solve the crash. That seems outside of our control. Even with more powerfull specs than my laptop, the crash still happened in the VM but didn't in my computer.

I'll turn the WSL1 test case off for now and park a card to investigate this later.